# docs/openapi/v1/openapi.yaml
...
openapi: 3.0.3
info:
  title: Liftr Resource API
  version: "1.0"
  description: |
    Provisioner-neutral resource lifecycle control plane.

    Authentication: every /v1 endpoint requires a bearer credential — an
    RFC 9068 JWT access token (typ at+jwt) issued by the configured issuer,
    carrying the Liftr API audience. Credential failures answer 401 with one
    identical Problem body; failure reasons are never disclosed. Health and
    readiness endpoints stay unauthenticated.

    Authorization: actions are authorized against Resource ownership. A
    create requires permission under the requested owner; reads, updates, and
    deletes require permission under the stored owner; Operations authorize
    through their owning Resource; ResourceType discovery is readable by every
    authenticated principal. Forbidden Resource access is externally
    indistinguishable from absence (404) so existence, generation, state, and
    outputs are never disclosed to unauthorized callers; forbidden creates
    answer 403 FORBIDDEN.

    Idempotency keys are scoped per authenticated principal: two principals
    may use the same key independently, and possessing another principal's
    key never grants access to its recorded result.

    Mutations are asynchronous admissions. POST creates a Resource and
    returns 201 immediately; PUT and DELETE return 202 while infrastructure
    work proceeds out of band. Clients monitor lifecycle progress through
    Operations.

...

...

...

...

...

...
security:
  - bearerAuth: []
paths:
  /v1/resource-types:
    get:
      operationId: listResourceTypes
      summary: List discoverable ResourceType contracts
      description: |
        Returns every registered ResourceType as a summary, ordered
        deterministically by name ascending, then version ascending.
        Summaries never embed the spec schema; fetch the detail endpoint for
        it. No pagination exists in v1 because the catalog is small and the
        order is total.
      parameters:
        - $ref: "#/components/parameters/XCorrelationId"
      responses:
        "200":
          description: The deterministic list of ResourceType summaries.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestId"
            Cache-Control:
              $ref: "#/components/headers/CacheControlNoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceTypeList"
              example:
                items:
                  - name: PostgreSQLDatabase
                    version: v1
                    displayName: PostgreSQL Database
                    description: A managed PostgreSQL database requested through a provisioner-neutral contract.
                    capabilities: [create, delete, observe, update]
                    href: /v1/resource-types/PostgreSQLDatabase/v1
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/Internal"
        "503":
          $ref: "#/components/responses/PersistenceUnavailable"
  /v1/resource-types/{name}/{version}:
    parameters:
      - $ref: "#/components/parameters/ResourceTypeName"
      - $ref: "#/components/parameters/ResourceTypeVersion"
    get:
      operationId: getResourceType
      summary: Read one ResourceType contract with its spec schema
      description: |
        Returns one developer contract including its embedded "specSchema".
        The schema document follows JSON Schema draft 2020-12; this API models
        only its envelope and does not own the schema keywords. The "$id" URN
        identifies the schema and is never fetched by Liftr or clients.
      parameters:
        - $ref: "#/components/parameters/XCorrelationId"
      responses:
        "200":
          description: The detailed ResourceType representation.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestId"
            Cache-Control:
              $ref: "#/components/headers/CacheControlNoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceType"
        "404":
          $ref: "#/components/responses/ResourceTypeNotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/Internal"
        "503":
          $ref: "#/components/responses/PersistenceUnavailable"
...
  /v1/operations/{id}:
    parameters:
      - $ref: "#/components/parameters/OperationId"
    get:
      operationId: getOperation
      summary: Read one lifecycle Operation
      description: |
        Reports lifecycle state without internal execution detail. This
        response does not represent a Resource and therefore carries no
        Liftr-Generation header.
      parameters:
        - $ref: "#/components/parameters/XCorrelationId"
      responses:
        "200":
          description: The Operation representation.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestId"
            Cache-Control:
              $ref: "#/components/headers/CacheControlNoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        "404":
          description: OPERATION_NOT_FOUND.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                operationNotFound:
                  value:
                    type: https://liftr.dev/problems/operation-not-found
                    title: Operation not found
                    status: 404
                    instance: /v1/operations/op-example
                    code: OPERATION_NOT_FOUND
                    requestId: 6f2c7a4c8e5b41d0
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/Internal"
        "503":
          $ref: "#/components/responses/PersistenceUnavailable"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        RFC 9068 JWT access token (typ at+jwt) issued by the configured
        issuer and carrying the Liftr API audience. Liftr verifies tokens;
        it never issues or proxies them.
  parameters:
...
    OperationId:
      name: id
      in: path
      required: true
      description: Lifecycle Operation identifier.
      schema:
        type: string
    ResourceTypeName:
      name: name
      in: path
      required: true
      description: ResourceType name, addressed as one URL segment.
      schema:
        type: string
    ResourceTypeVersion:
      name: version
      in: path
      required: true
      description: |
        Opaque ResourceType version. A name/version pair identifies one
        immutable contract; breaking and additive schema changes require a new
        version once Resources exist.
      schema:
        type: string
...
    XCorrelationId:
      name: X-Correlation-ID
      in: header
      required: false
      description: |
        Optional client correlation identifier. It stays strictly separate
        from the authoritative X-Request-ID and is echoed verbatim.
      schema:
        type: string
  headers:
...
    XRequestId:
      description: Authoritative server-generated request identifier.
      schema:
        type: string
    CacheControlNoStore:
      description: Versioned representations carry no caching metadata.
      schema:
        type: string
        enum: [no-store]
  schemas:
...
    Capability:
      type: string
      description: |
        A developer-facing action supported by a ResourceType contract.
        "observe" is contract metadata only; v1 defines no developer-facing
        observe endpoint.
      enum: [create, update, delete, observe]
    ResourceTypeSummary:
      type: object
      description: |
        Discovery summary of one developer contract. Capabilities are
        ResourceType-contract capabilities; they do not report provisioner
        selection, backend health, or current availability. Provisioner and
        platform concepts have no representation in discovery.
      required: [name, version, displayName, description, capabilities, href]
      properties:
        name:
          type: string
        version:
          type: string
        displayName:
          type: string
        description:
          type: string
        capabilities:
          type: array
          items:
            $ref: "#/components/schemas/Capability"
        href:
          type: string
          description: Relative URL of the detailed contract representation.
    # specSchema embeds a JSON Schema draft 2020-12 document exactly as the
    # ResourceType registered it. The core API models only this envelope;
    # keyword semantics are governed by the JSON Schema specification, which
    # OpenAPI 3.0 Schema Objects deliberately do not mirror here.
    ResourceSpecSchema:
      type: object
      additionalProperties: true
      description: |
        JSON Schema (draft 2020-12) describing the accepted "spec" shape for
        one ResourceType version. Schemas are self-contained; validation never
        performs network resolution. The "$id" URN identifies the schema and
        is never fetched. Structural constraints (types, required properties,
        enums, ranges) apply at admission; unknown properties are rejected.
    ResourceType:
      type: object
      required: [name, version, displayName, description, capabilities, href, specSchema]
      properties:
        name:
          type: string
        version:
          type: string
        displayName:
          type: string
        description:
          type: string
        capabilities:
          type: array
          items:
            $ref: "#/components/schemas/Capability"
        href:
          type: string
        specSchema:
          $ref: "#/components/schemas/ResourceSpecSchema"
        outputContract:
          $ref: "#/components/schemas/ResourceTypeOutputContract"
    SpecViolation:
      type: object
      description: |
        One sanitized spec-contract violation. Paths are RFC 6901 JSON
        Pointers into the submitted spec ("" denotes the root). Violations are
        deterministically ordered and capped at ten entries per response.
      required: [path, keyword, message]
      properties:
        path:
          type: string
        keyword:
          type: string
        message:
          type: string
    ResourceTypeList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ResourceTypeSummary"
    ResourceTypeOutputField:
      type: object
      description: |
        One declared developer-consumable output of a ResourceType version.
        Names are provider-neutral and never echo implementation identifiers;
        secret material has no representation in v1.
      required: [name, jsonType, requiredWhenReady]
      properties:
        name:
          type: string
        jsonType:
          type: string
          description: The exact scalar JSON type of this output value.
          enum: [string, integer, number, boolean]
        requiredWhenReady:
          type: boolean
          description: |
            When true, a successfully reconciled generation must publish this
            field before Liftr reports reconciliation success.
    ResourceTypeOutputContract:
      type: object
      description: |
        The declared non-secret output contract of one ResourceType version.
        It is part of the immutable contract identity; the core API models
        only this envelope, while each contract declares its exact fields.
      required: [fields]
      properties:
        fields:
          type: array
          items:
            $ref: "#/components/schemas/ResourceTypeOutputField"
...
    OperationFailure:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
        message:
          type: string
    Operation:
      type: object
      description: |
        One asynchronous action against a specific desired-state generation.
        Internal execution phases have no public representation in v1.
      required:
        - id
        - resourceId
        - capability
        - state
        - targetGeneration
        - requestedAt
      properties:
        id:
          type: string
        resourceId:
          type: string
        capability:
          type: string
          enum: [create, update, delete]
        state:
          type: string
          enum: [Pending, Running, Succeeded, Failed, Canceled]
        targetGeneration:
          type: integer
          format: int64
        requestedAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        failure:
          $ref: "#/components/schemas/OperationFailure"
    Problem:
      type: object
      description: RFC 9457 Problem Details with Liftr extensions.
      required: [type, title, status, code, requestId]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
          description: Stable machine-readable Liftr error code.
          enum:
            - UNAUTHENTICATED
            - FORBIDDEN
            - INVALID_ARGUMENT
            - UNSUPPORTED_RESOURCE_TYPE
            - RESOURCE_TYPE_NOT_FOUND
            - RESOURCE_SPEC_INVALID
            - RESOURCE_NOT_FOUND
            - OPERATION_NOT_FOUND
            - RESOURCE_ALREADY_EXISTS
            - IDEMPOTENCY_CONFLICT
            - GENERATION_CONFLICT
            - OPERATION_ACTIVE
            - RESOURCE_STATE_CONFLICT
            - UNSUPPORTED_CAPABILITY
            - PRECONDITION_REQUIRED
            - PROVISIONER_UNAVAILABLE
            - PERSISTENCE_UNAVAILABLE
            - INTERNAL
        requestId:
          type: string
        currentGeneration:
          type: integer
          format: int64
          description: Narrow extension present on GENERATION_CONFLICT problems.
        violations:
          type: array
          description: |
            Present on RESOURCE_SPEC_INVALID problems. Sanitized, deterministically
            ordered spec-contract violations; capped at ten entries.
          items:
            $ref: "#/components/schemas/SpecViolation"
        truncated:
          type: boolean
          description: |
            True when RESOURCE_SPEC_INVALID had more violations than the
            response carries; absent otherwise.
  responses:
    Unauthorized:
      description: |
        UNAUTHENTICATED; no usable credential was presented or the presented
        credential failed verification. The body is identical in every case.
        WWW-Authenticate carries error="invalid_token" only when a credential
        was presented.
      headers:
        WWW-Authenticate:
          description: Bearer challenge; see RFC 6750.
          schema:
            type: string
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: https://liftr.dev/problems/unauthenticated
            title: Unauthenticated
            status: 401
            detail: valid bearer credentials are required
            code: UNAUTHENTICATED
            requestId: 6f2c7a4c8e5b41d0
...
    ResourceTypeNotFound:
      description: RESOURCE_TYPE_NOT_FOUND; no contract is registered with this name and version.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: https://liftr.dev/problems/resource-type-not-found
            title: Resource type not found
            status: 404
            instance: /v1/resource-types/NoDatabase/v1
            code: RESOURCE_TYPE_NOT_FOUND
            requestId: 6f2c7a4c8e5b41d0
...
    PersistenceUnavailable:
      description: PERSISTENCE_UNAVAILABLE.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Internal:
      description: INTERNAL; unexpected server-side failure without internal detail.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"